autoparser: move parse_choices(), read_data() to util
